Add JPA auditing, enhance error handling, and update listing request/response structures

+++ src/main/java/com/evarkdasim/evarkadasim_api/repository/CityRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.math.BigInteger;
 
-public interface CityRepository extends JpaRepository<City, BigInteger> {
+public interface CityRepository extends JpaRepository<City, Long> {
 
 
 }
