+++ src/main/java/com/evarkdasim/evarkadasim_api/repository/NeighborhoodRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.math.BigInteger;
 
-public interface NeighborhoodRepository extends JpaRepository<Neighborhood, BigInteger> {
+public interface NeighborhoodRepository extends JpaRepository<Neighborhood, Long> {
 }
